[clients] remove keyboard state sync on CAPS release
The keyboard state was synced in xfreerdp (and more recently
sdl3-freerdp) when the CAPS key was released.
Some investigation revealed this to be a historic workaround for loss of
keyboard state sync (e.g. NUK, CAPS or SCROLL getting out of sync)

This can not happen anymore as all events during windows focus are
processed properly and every time the keyboard state changes outside it
is synchronized when focus is gained.