kerberos: do various tries for TGT retrieval in u2u
In a server-side user2user scenario, you can have either a configuration with
a service mapped to a user or something host based. So your keytab will contain
either a TERMSRV/<host>@<realm> entry that is able to retrieve a TGT, or the
machine account associated with <host>$@<realm>.

This patch makes the kerberos SSPI first try to retrieved a TGT as if it was
a service mapped to a user. And if it doesn't work (or if the entry is missing),
it goes for a host based attempt.